Issue #284: Standardize chart save/export filenames between Dashboard and Lite
Add chart right-click context menu to all 21 Lite charts (Copy Image, Save Image As,
Open in New Window, Revert, Export CSV) with descriptive filenames using yyyy-MM-dd_HH-mm-ss
timestamp format. Consolidate Dashboard's duplicate SetupChartSaveMenu into the shared
TabHelpers.SetupChartContextMenu, and add context menus to the 2 Current Waits charts
that had none.
